requirements.txt
python-telegram-bot
requests
pandas
beautifulsoup4
whois
openpyxl